Fixed memory leaks in task launcher.
Fixed memory leaks when tas launcher is reloaded.

Fix crash in TidyScrollBar when it is disposed.
Fix access to member data->cancelled after it data is freed in HdLauncherTree.
Destroy old HdLauncherPages when the tree is reloaded in HDLauncher.
Do not leak TidyAdjustments in HdLauncherGrid.
Remove unused code and leaks in HdLauncherPage.