reverting c9e5a7423b59a88b7fdd836c296eaf569e5ee60e
because it caused a regression: be in tasw and have an app open an
appmodal dialog.  hd is supposed to stay in tasw and show the dialog
popping up, but with this patch it switched to APP state (and the
actors were all hidden because tasw hides everything it owned in
the hope hdrm will show whatever is necessary for it).