Try to do the right thing on LAUNCHER's exit
ON TL in potrait mode exit, it should rotate in landscape when the next state
do not support PT, or stay in PT then it does.

Currently goint from TL to:

HOME ('back'): it show the first frame of what appears a zoom in (like the one on
TRANSITION_OUT from LAUNCHER to HOME in landscape). My guess is that the zoom
in transition is stopped to make the rotation start, but the first frame is
shown anyway.

LOADING: the background of the next app appears partially, as it was oriente
in the wrong direction. ie: from portrait mode launch 'photos' and take a look
to the black background.