Single- and two-column task switcher (configurable via transitions.ini)
This adds a new configuration option "taskswitcher" to the
transitions.ini file that controls the way the tasks are
laid out in the task switcher. Three options are provided:

  0 ... the default Maemo 5 layout
  1 ... single-column layout
  2 ... two-column layout

The changes are applied at runtime, but only when the layout
is changed (i.e. when a window is closed or opened).